Use mapfile to implement the file system offload
