+++ conf/broker.conf
@@ -712,17 +712,8 @@ gcsManagedLedgerOffloadReadBufferSizeInBytes=1048576
 # For more details, see the "Service Accounts" section of https://support.google.com/googleapi/answer/6158849
 gcsManagedLedgerOffloadServiceAccountKeyFile=
 
-#The base path for storage
-hdfsFileSystemManagedLedgerOffloadStorageBasePath=
-
-#The user name for file system
-hdfsFileSystemManagedLedgerOffloadUserName=
-
-#The buffer size of the inputStream for reading entry
-hdfsFileSystemReadHandleReadBufferSize=
-
-#The uri of file system
-hdfsFileSystemManagedLedgerOffloadAccessUri=
+#For File System Storage, file system profile path
+fileSystemProfilePath=../conf/core-site.xml
 
 ### --- Deprecated config variables --- ###
