+++ tiered-storage/file-system/src/main/java/org/apache/bookkeeper/mledger/offload/filesystem/FileSystemLedgerOffloaderFactory.java
@@ -35,6 +35,6 @@ public class FileSystemLedgerOffloaderFactory implements LedgerOffloaderFactory<
     @Override
     public FileSystemManagedLedgerOffloader create(Properties properties, Map<String, String> userMetadata, OrderedScheduler scheduler) throws IOException {
         TieredStorageConfigurationData data = TieredStorageConfigurationData.create(properties);
-        return FileSystemManagedLedgerOffloader.create(data, userMetadata, scheduler);
+        return FileSystemManagedLedgerOffloader.create(data, scheduler);
     }
 }
