--- tiered-storage/file-system/src/main/java/org/apache/bookkeeper/mledger/offload/filesystem/OffloadIndex.java
@@ -1,65 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.bookkeeper.mledger.offload.filesystem;
-
-import org.apache.bookkeeper.client.api.LedgerMetadata;
-import org.apache.bookkeeper.common.annotation.InterfaceStability.Unstable;
-
-import java.io.Closeable;
-import java.io.IOException;
-
-/**
- *
- * The Index block abstraction used for offload a ledger to long term storage.
- *
- */
-@Unstable
-public interface OffloadIndex extends Closeable {
-
-    /**
-     * Get the related OffloadIndexEntry that contains the given messageEntryId.
-     *
-     * @param entryId
-     *                      the entry id of message
-     * @return the offload index entry
-     */
-    OffloadIndexEntry getFloorIndexEntryByEntryId(long entryId) throws IOException;
-
-    /**
-     * Get the entry count that contained in this index Block.
-     */
-    int getEntryCount();
-
-    /**
-     * Get LedgerMetadata.
-     */
-    LedgerMetadata getLedgerMetadata();
-
-    /**
-     * Get the total size of the data object.
-     */
-    long getDataObjectLength();
-
-    /**
-     * Get the length of the header in the blocks in the data object.
-     */
-    long getDataHeaderLength();
-
-}
-
