+++ tiered-storage/file-system/src/main/java/org/apache/bookkeeper/mledger/offload/filesystem/impl/FileStoreBackedReadHandleImpl.java
@@ -18,9 +18,11 @@
  */
 package org.apache.bookkeeper.mledger.offload.filesystem.impl;
 
+import com.google.common.collect.Maps;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.PooledByteBufAllocator;
 import org.apache.bookkeeper.client.BKException;
+import org.apache.bookkeeper.client.api.DigestType;
 import org.apache.bookkeeper.client.api.LastConfirmedAndEntry;
 import org.apache.bookkeeper.client.api.LedgerEntries;
 import org.apache.bookkeeper.client.api.LedgerEntry;
@@ -29,40 +31,46 @@ import org.apache.bookkeeper.client.api.ReadHandle;
 import org.apache.bookkeeper.client.impl.LedgerEntriesImpl;
 import org.apache.bookkeeper.client.impl.LedgerEntryImpl;
 
-import org.apache.bookkeeper.mledger.offload.filesystem.FileSystemReadHandleInputStream;
-import org.apache.bookkeeper.mledger.offload.filesystem.OffloadIndexEntry;
-import org.apache.bookkeeper.mledger.offload.filesystem.OffloadIndexFile;
-import org.apache.hadoop.fs.FSDataInputStream;
-import org.apache.hadoop.fs.FileSystem;
-import org.apache.hadoop.fs.Path;
+import org.apache.bookkeeper.net.BookieSocketAddress;
+import org.apache.bookkeeper.proto.DataFormats;
+import org.apache.hadoop.io.BytesWritable;
+import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.MapFile;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.DataInputStream;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.NavigableMap;
+import java.util.TreeMap;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.ScheduledExecutorService;
 
 public class FileStoreBackedReadHandleImpl implements ReadHandle {
     private static final Logger log = LoggerFactory.getLogger(FileStoreBackedReadHandleImpl.class);
-    private final long ledgerId;
     private final ExecutorService executor;
-    private final OffloadIndexFile offloadIndexFile;
-    private final DataInputStream dataInputStream;
-    private final FileSystemReadHandleInputStream inputStream;
+    private final MapFile.Reader reader;
+    private final long ledgerId;
+    private final LedgerMetadata ledgerMetadata;
 
-    private FileStoreBackedReadHandleImpl(ExecutorService executor, long ledgerId,
-                                          OffloadIndexFile offloadIndexFile, FSDataInputStream fsDataInputStream, int readBufferSize) {
+    private FileStoreBackedReadHandleImpl(ExecutorService executor, MapFile.Reader reader, long ledgerId) throws IOException {
         this.ledgerId = ledgerId;
         this.executor = executor;
-        this.offloadIndexFile = offloadIndexFile;
-        this.inputStream = new FileSystemReadHandleInputStreamImpl(fsDataInputStream,
-                readBufferSize, offloadIndexFile.getDataObjectLength(),
-                offloadIndexFile.getDataHeaderLength());
-        this.dataInputStream = new DataInputStream(this.inputStream);
+        this.reader = reader;
+        LongWritable key = new LongWritable();
+        BytesWritable value = new BytesWritable();
+        try {
+            key.set(-1);
+            reader.get(key, value);
+            this.ledgerMetadata = parseLedgerMetadata(value.copyBytes());
+        } catch (IOException e) {
+            log.error("Fail to read LedgerMetadata for key {}",
+                    key.get());
+            throw new IOException("Fail to read LedgerMetadata for key " + key.get());
+        }
     }
 
     @Override
@@ -72,7 +80,8 @@ public class FileStoreBackedReadHandleImpl implements ReadHandle {
 
     @Override
     public LedgerMetadata getLedgerMetadata() {
-        return offloadIndexFile.getLedgerMetadata();
+        return ledgerMetadata;
+
     }
 
     @Override
@@ -80,9 +89,7 @@ public class FileStoreBackedReadHandleImpl implements ReadHandle {
         CompletableFuture<Void> promise = new CompletableFuture<>();
         executor.submit(() -> {
                 try {
-                    dataInputStream.close();
-                    inputStream.close();
-                    promise.complete(null);
+                    reader.close();
                 } catch (IOException t) {
                     promise.completeExceptionally(t);
                 }
@@ -95,41 +102,42 @@ public class FileStoreBackedReadHandleImpl implements ReadHandle {
         log.debug("Ledger {}: reading {} - {}", getId(), firstEntry, lastEntry);
         CompletableFuture<LedgerEntries> promise = new CompletableFuture<>();
         executor.submit(() -> {
-                if (firstEntry > lastEntry
+            if (firstEntry > lastEntry
                     || firstEntry < 0
                     || lastEntry > getLastAddConfirmed()) {
-                    promise.completeExceptionally(new BKException.BKIncorrectParameterException());
-                    return;
-                }
-                long entriesToRead = (lastEntry - firstEntry) + 1;
-                List<LedgerEntry> entries = new ArrayList<LedgerEntry>();
-                long nextExpectedId = firstEntry;
-                try {
-                    OffloadIndexEntry floorEntry = offloadIndexFile.getFloorIndexEntryByEntryId(firstEntry);
-                    inputStream.seek(floorEntry.getOffset());
-                    while (entriesToRead > 0) {
-                        int length = dataInputStream.readInt();
-                        long entryId = dataInputStream.readLong();
-                        if (entryId == nextExpectedId) {
-                            ByteBuf entryBuf = PooledByteBufAllocator.DEFAULT.buffer(length, length);
-                            entries.add(LedgerEntryImpl.create(ledgerId, entryId, length, entryBuf));
-                            inputStream.readByteBuf(entryBuf);
-                            entriesToRead--;
-                            nextExpectedId++;
-                        } else if (entryId > lastEntry) {
-                            log.info("Expected to read {}, but read {}, which is greater than last entry {}",
-                                    nextExpectedId, entryId, lastEntry);
-                            throw new BKException.BKUnexpectedConditionException();
-                        } else {
-                            inputStream.skipLength(length);
-                        }
+                promise.completeExceptionally(new BKException.BKIncorrectParameterException());
+                return;
+            }
+            long entriesToRead = (lastEntry - firstEntry) + 1;
+            List<LedgerEntry> entries = new ArrayList<LedgerEntry>();
+            long nextExpectedId = firstEntry;
+            LongWritable key = new LongWritable();
+            BytesWritable value = new BytesWritable();
+            try {
+                key.set(nextExpectedId - 1);
+                reader.seek(key);
+                while (entriesToRead > 0) {
+                    reader.next(key, value);
+                    int length = value.getLength();
+                    long entryId = key.get();
+                    if (entryId == nextExpectedId) {
+                        ByteBuf buf = PooledByteBufAllocator.DEFAULT.buffer(length, length);
+                        entries.add(LedgerEntryImpl.create(ledgerId, entryId, length, buf));
+                        buf.writeBytes(value.copyBytes());
+                        entriesToRead--;
+                        nextExpectedId++;
+                    } else if (entryId > lastEntry) {
+                        log.info("Expected to read {}, but read {}, which is greater than last entry {}",
+                                nextExpectedId, entryId, lastEntry);
+                        throw new BKException.BKUnexpectedConditionException();
                     }
-                    promise.complete(LedgerEntriesImpl.create(entries));
-                } catch (Throwable t) {
-                    promise.completeExceptionally(t);
-                    entries.forEach(LedgerEntry::close);
-                }
-            });
+            }
+                promise.complete(LedgerEntriesImpl.create(entries));
+            } catch (Throwable t) {
+                promise.completeExceptionally(t);
+                entries.forEach(LedgerEntry::close);
+            }
+        });
         return promise;
     }
 
@@ -172,12 +180,142 @@ public class FileStoreBackedReadHandleImpl implements ReadHandle {
         return promise;
     }
 
-    public static ReadHandle open(ScheduledExecutorService executor, String dataFilePath,
-                                  String indexFilePath, FileSystem fileSystem, long ledgerId, int readBufferSize) throws IOException {
-            OffloadIndexFile offloadIndexFile =
-                    OffloadIndexFileImpl.getRecycler().get()
-                            .initIndexFile(fileSystem.open(new Path(indexFilePath)));
-            return new FileStoreBackedReadHandleImpl(executor, ledgerId,
-                    offloadIndexFile, fileSystem.open(new Path(dataFilePath)), readBufferSize);
+    public static ReadHandle open(ScheduledExecutorService executor, MapFile.Reader reader, long ledgerId) throws IOException {
+            return new FileStoreBackedReadHandleImpl(executor, reader, ledgerId);
     }
+
+    static private class InternalLedgerMetadata implements LedgerMetadata {
+        private DataFormats.LedgerMetadataFormat ledgerMetadataFormat;
+
+        private int ensembleSize;
+        private int writeQuorumSize;
+        private int ackQuorumSize;
+        private long lastEntryId;
+        private long length;
+        private DataFormats.LedgerMetadataFormat.DigestType digestType;
+        private long ctime;
+        private byte[] password;
+        private State state;
+        private Map<String, byte[]> customMetadata = Maps.newHashMap();
+        private TreeMap<Long, ArrayList<BookieSocketAddress>> ensembles = new TreeMap<Long, ArrayList<BookieSocketAddress>>();
+
+        InternalLedgerMetadata(DataFormats.LedgerMetadataFormat ledgerMetadataFormat) {
+            this.ensembleSize = ledgerMetadataFormat.getEnsembleSize();
+            this.writeQuorumSize = ledgerMetadataFormat.getQuorumSize();
+            this.ackQuorumSize = ledgerMetadataFormat.getAckQuorumSize();
+            this.lastEntryId = ledgerMetadataFormat.getLastEntryId();
+            this.length = ledgerMetadataFormat.getLength();
+            this.digestType = ledgerMetadataFormat.getDigestType();
+            this.ctime = ledgerMetadataFormat.getCtime();
+            this.state = State.CLOSED;
+            this.password = ledgerMetadataFormat.getPassword().toByteArray();
+
+            if (ledgerMetadataFormat.getCustomMetadataCount() > 0) {
+                ledgerMetadataFormat.getCustomMetadataList().forEach(
+                        entry -> this.customMetadata.put(entry.getKey(), entry.getValue().toByteArray()));
+            }
+
+            ledgerMetadataFormat.getSegmentList().forEach(segment -> {
+                ArrayList<BookieSocketAddress> addressArrayList = new ArrayList<BookieSocketAddress>();
+                segment.getEnsembleMemberList().forEach(address -> {
+                    try {
+                        addressArrayList.add(new BookieSocketAddress(address));
+                    } catch (IOException e) {
+                        log.error("Exception when create BookieSocketAddress. ", e);
+                    }
+                });
+                this.ensembles.put(segment.getFirstEntryId(), addressArrayList);
+            });
+        }
+
+        @Override
+        public boolean hasPassword() { return true; }
+
+        @Override
+        public byte[] getPassword() { return password; }
+
+        @Override
+        public State getState() { return state; }
+
+        @Override
+        public int getMetadataFormatVersion() { return 2; }
+
+        @Override
+        public int getEnsembleSize() {
+            return this.ensembleSize;
+        }
+
+        @Override
+        public int getWriteQuorumSize() {
+            return this.writeQuorumSize;
+        }
+
+        @Override
+        public int getAckQuorumSize() {
+            return this.ackQuorumSize;
+        }
+
+        @Override
+        public long getLastEntryId() {
+            return this.lastEntryId;
+        }
+
+        @Override
+        public long getLength() {
+            return this.length;
+        }
+
+        @Override
+        public DigestType getDigestType() {
+            switch (this.digestType) {
+                case HMAC:
+                    return DigestType.MAC;
+                case CRC32:
+                    return DigestType.CRC32;
+                case CRC32C:
+                    return DigestType.CRC32C;
+                case DUMMY:
+                    return DigestType.DUMMY;
+                default:
+                    throw new IllegalArgumentException("Unable to convert digest type " + digestType);
+            }
+        }
+
+        @Override
+        public long getCtime() {
+            return this.ctime;
+        }
+
+        @Override
+        public boolean isClosed() {
+            return this.state == State.CLOSED;
+        }
+
+        @Override
+        public Map<String, byte[]> getCustomMetadata() {
+            return this.customMetadata;
+        }
+
+        @Override
+        public List<BookieSocketAddress> getEnsembleAt(long entryId) {
+            return ensembles.get(ensembles.headMap(entryId + 1).lastKey());
+        }
+
+        @Override
+        public NavigableMap<Long, ? extends List<BookieSocketAddress>> getAllEnsembles() {
+            return this.ensembles;
+        }
+
+        @Override
+        public String toSafeString() {
+            return toString();
+        }
+    }
+
+    private static LedgerMetadata parseLedgerMetadata(byte[] bytes) throws IOException {
+        DataFormats.LedgerMetadataFormat.Builder builder = DataFormats.LedgerMetadataFormat.newBuilder();
+        builder.mergeFrom(bytes);
+        return new InternalLedgerMetadata(builder.build());
+    }
+
 }
