+++ tiered-storage/file-system/src/test/java/org/apache/bookkeeper/mledger/offload/filesystem/impl/FileSystemManagedLedgerOffloaderTest.java
@@ -29,10 +29,8 @@ import org.apache.bookkeeper.client.api.LedgerEntry;
 import org.apache.bookkeeper.client.api.ReadHandle;
 import org.apache.bookkeeper.mledger.LedgerOffloader;
 import org.apache.bookkeeper.mledger.offload.filesystem.FileStoreTestBase;
-import org.apache.bookkeeper.mledger.offload.filesystem.FileSystemEntryBytesReader;
 import org.apache.bookkeeper.util.ZkUtils;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.zookeeper.CreateMode;
@@ -53,12 +51,10 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 
 public class FileSystemManagedLedgerOffloaderTest extends FileStoreTestBase {
     private final PulsarMockBookKeeper bk;
-    private long objectLength = FileSystemEntryBytesReader.getDataHeaderLength();
     private String topic = "public/default/persistent/testOffload";
     private String storagePath = createStoragePath(topic);
     private LedgerHandle lh;
     private ReadHandle toWrite;
-    private long idOfExceedReadBuffer;
     private final int numberOfEntries = 600;
     private  Map<String, String> map = new HashMap<>();
 
@@ -88,11 +84,6 @@ public class FileSystemManagedLedgerOffloaderTest extends FileStoreTestBase {
         int blocksWritten = 1;
         while (blocksWritten <= numberOfEntries) {
             byte[] entry = ("foobar"+i).getBytes();
-            objectLength += entry.length + FileSystemEntryBytesReader.getEntryHeaderSize();
-            if (objectLength > getReadBufferSize() + FileSystemEntryBytesReader.getDataHeaderLength()) {
-                idOfExceedReadBuffer = lh.readLastEntry().getEntryId();
-            }
-
             blocksWritten++;
             lh.addEntry(entry);
             i++;
@@ -102,16 +93,7 @@ public class FileSystemManagedLedgerOffloaderTest extends FileStoreTestBase {
         return bk.newOpenLedgerOp().withLedgerId(lh.getId())
                 .withPassword("foobar".getBytes()).withDigestType(DigestType.CRC32).execute().get();
     }
-    @Test
-    public void testOffload() throws Exception {
-        LedgerOffloader offloader = fileSystemManagedLedgerOffloader;
-        UUID uuid = UUID.randomUUID();
-        offloader.offload(toWrite, uuid, map).get();
-        Configuration configuration = new Configuration();
-        FileSystem fileSystem = FileSystem.get(new URI(getURI()), configuration);
-        FileStatus fileStatus = fileSystem.getFileStatus(new Path(createDataFilePath(storagePath, lh.getId(), uuid)));
-        Assert.assertEquals(objectLength, fileStatus.getLen());
-    }
+
     @Test
     public void testOffloadAndRead() throws Exception {
         LedgerOffloader offloader = fileSystemManagedLedgerOffloader;
@@ -119,8 +101,8 @@ public class FileSystemManagedLedgerOffloaderTest extends FileStoreTestBase {
         offloader.offload(toWrite, uuid, map).get();
         ReadHandle toTest = offloader.readOffloaded(toWrite.getId(), uuid, map).get();
         Assert.assertEquals(toTest.getLastAddConfirmed(), toWrite.getLastAddConfirmed());
-        LedgerEntries toTestEntries = toTest.read(0, idOfExceedReadBuffer - 1);
-        LedgerEntries toWriteEntries = toWrite.read(0,idOfExceedReadBuffer - 1);
+        LedgerEntries toTestEntries = toTest.read(0, numberOfEntries - 1);
+        LedgerEntries toWriteEntries = toWrite.read(0,numberOfEntries - 1);
         Iterator<LedgerEntry> toTestIter = toTestEntries.iterator();
         Iterator<LedgerEntry> toWriteIter = toWriteEntries.iterator();
         while(toTestIter.hasNext()) {
@@ -132,21 +114,8 @@ public class FileSystemManagedLedgerOffloaderTest extends FileStoreTestBase {
             Assert.assertEquals(toWriteEntry.getLength(), toTestEntry.getLength());
             Assert.assertEquals(toWriteEntry.getEntryBuffer(), toTestEntry.getEntryBuffer());
         }
-        toTestEntries = toTest.read(0, idOfExceedReadBuffer);
-        toWriteEntries = toWrite.read(0,idOfExceedReadBuffer);
-        toTestIter = toTestEntries.iterator();
-        toWriteIter = toWriteEntries.iterator();
-        while(toTestIter.hasNext()) {
-            LedgerEntry toWriteEntry = toWriteIter.next();
-            LedgerEntry toTestEntry = toTestIter.next();
-
-            Assert.assertEquals(toWriteEntry.getLedgerId(), toTestEntry.getLedgerId());
-            Assert.assertEquals(toWriteEntry.getEntryId(), toTestEntry.getEntryId());
-            Assert.assertEquals(toWriteEntry.getLength(), toTestEntry.getLength());
-            Assert.assertEquals(toWriteEntry.getEntryBuffer(), toTestEntry.getEntryBuffer());
-        }
-        toTestEntries = toTest.read(0, numberOfEntries - 1);
-        toWriteEntries = toWrite.read(0, numberOfEntries - 1);
+        toTestEntries = toTest.read(1, numberOfEntries - 1);
+        toWriteEntries = toWrite.read(1,numberOfEntries - 1);
         toTestIter = toTestEntries.iterator();
         toWriteIter = toWriteEntries.iterator();
         while(toTestIter.hasNext()) {
@@ -159,8 +128,9 @@ public class FileSystemManagedLedgerOffloaderTest extends FileStoreTestBase {
             Assert.assertEquals(toWriteEntry.getEntryBuffer(), toTestEntry.getEntryBuffer());
         }
     }
+
     @Test
-    public void testRatherThanWrite() throws Exception {
+    public void testDeleteOffload() throws Exception {
         LedgerOffloader offloader = fileSystemManagedLedgerOffloader;
         UUID uuid = UUID.randomUUID();
         offloader.offload(toWrite, uuid, map).get();
@@ -173,27 +143,15 @@ public class FileSystemManagedLedgerOffloaderTest extends FileStoreTestBase {
         Assert.assertEquals(false, fileSystem.exists(new Path(createIndexFilePath(storagePath, lh.getId(), uuid))));
     }
 
-    @Test
-    public void testDeleteOffload() throws Exception {
-        LedgerOffloader offloader = fileSystemManagedLedgerOffloader;
-        UUID uuid = UUID.randomUUID();
-        Map<String, String> map = new HashMap<>();
-        map.put("ManagedLedgerName", topic);
-        offloader.offload(toWrite, uuid, map).get();
-        ReadHandle toTest = offloader.readOffloaded(toWrite.getId(), uuid, map).get();
-
-        offloader.deleteOffloaded(lh.getId(), uuid, map);
-    }
-
     private String createStoragePath(String managedLedgerName) {
         return basePath + "/" + managedLedgerName + "/";
     }
 
     private String createIndexFilePath(String storagePath, long ledgerId, UUID uuid) {
-        return storagePath + ledgerId + "-" + uuid + ".index";
+        return storagePath + ledgerId + "-" + uuid;
     }
 
     private String createDataFilePath(String storagePath, long ledgerId, UUID uuid) {
-        return storagePath + ledgerId + "-" + uuid + ".log";
+        return storagePath + ledgerId + "-" + uuid;
     }
 }
